messeages page responsive for mobile devices
