--- src/pages/messenger/messenger.css
@@ -1,75 +1,0 @@
-.messenger{
-    display: flex;
-    height: calc(100vh - 70px);
-}
-
-.chatMenu{
-    flex: 3.5; 
-}
-
-.chatMenuSearch{
-    width: 90%;
-    padding: 10px 0;
-    border: none;
-    border-bottom: 1px solid gray;
-}
-
-.chatMenuSearch:focus{
-    outline: none;
-}
-
-.chatBox{
-    flex: 5.5;
-}
-
-.chatBoxContainer{
-    display: flex;
-    flex-direction: column;
-    position: relative;
-}
-
-.chatBoxTop{
-    height: 100%;
-    overflow-y: scroll;
-    padding-right: 10px;
-}
-
-.chatBoxBottom{
-    margin-top: 5px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-}
-
-.chatBoxBottomMessage{
-    width: 80%;
-    height: 90px;
-    padding: 10px;
-}
-
-.chatBoxBottomButton{
-    width: 70px;
-    height: 40px;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    background-color: #306130;
-    color: white;
-}
-
-.noConversation{
-    position: absolute;
-    top: 15%;
-    font-size: 50px;
-    color: rgb(202, 202, 202);
-    cursor: default;
-}
-
-.chatOnline{
-    flex: 3;
-}
-
-.chatMenuContainer, .chatBoxContainer, .chatOnlineContainer{
-    padding: 10px;
-    height: 100%;
-}
