+++ tailwind.config.js
@@ -6,6 +6,7 @@ module.exports = {
       colors: {
         customPrimary: "rgb(146, 239, 146)",
         textSecondary: "#f0f2f5",
+        textTertiary: "rgb(202, 202, 202)",
       },
       boxShadow: {
         custom: "0px 0px 18px -8px rgba(0, 0, 0, 0.37)",
